Skyline: merge from https://svn.code.sf.net/p/proteowizard/code/branches/work/20170320_smlibs - Changes to improve small molecule support in Skyline including support for multiple adducts per molecule, and spectral libraries.
FOR USERS
Skyline users interested in proteomics should notice few differences - just one or two more buttons or tabs that say "small molecules", which can be ignored for proteomics purposes.
Skyline users interested in small molecules should notice many new things:
Skyline's internal model for small molecules has been aligned with that of peptides - in the Targets tree now contains neutral molecules with children that have adducts. The ability to declare multiple adducts for a single molecule (an extension of Skyline's support for multiple protonated states for peptides) greatly extends the number of Skyline features that can be used with small molecules.
This also allows the use of spectral libraries for small molecules - NIST .msf format is the first to be supported, with more on the way. This means you can use the NIST tools to filter out a desired subset of molecules and adducts, and import that into Skyline as a library and use it to create a target list. The "n" best ranked peaks are used, where n is taken as the value from the Transition Settings Library tab "Pick: __ Product ions" setting.
The Skyline UI is increasingly aware of the nature of the document in use:
 - the popup menus for the various graph types no longer offer A, B, C, X, Y, Z, and Precursor as ion types when working with small molecules, but instead offer Fragments and Precursor (we use the catch-all "Fragments" since we don't have an inherent sense of fragment type for generalized small molecules).
- the Transition Settings Filter tab now has a sub-tab just for small molecule settings: you will notice that there are adduct lists for precursor and fragment filtering similar to the charge state lists for peptides. Note that you can pick adducts from a popup list, but you can always compose your own as well. Typically fragment ions are known by mass and charge only, so adducts like [M+] and [M-] are suitable here.
- the Library Explorer window now says "Molecule" instead of "Peptide" as appropriate for the library in use.
- the transition list reader has new columns for importing InChiKey, InChi, CAS and HMDB identifiers.

FOR DEVELOPERS
SrmDocument has a new property "DocumentType" which gives instantaneous information about whether a document is proteomic, small_molecule, or mixed. Empty docs report proteomic.
Peptide sequences passed around as strings have been replaced with the Target class, which abstracts the idea of peptide vs small molecule.
Charge as integer has been replaced with the Adduct class - where we use to say "charge=2" what we really meant was "doubly protonated", so we now internally use an Adduct "M+2H". (The proteomics UI still presents integers, though)
Adducts should be treated as non-nullable - use and depend on Adduct.EMPTY to mean "no adduct".
A fair bit of renaming has gone on in the code to make a clearer distinction between a small molecule and a small molecule ion. TransitionGroups have molecules, while Transitions have ions (which are molecules with adducts). All ions are molecules but not all molecules are ions. 
The code that converts proteomic documents to small molecule documents for test purposes now converts associated .blib files as well.
The .blib and its cache format have shifted to accommodate small molecules. The associated LibKey class has also changed to accomodate generalized ionization, where we can no longer assume integer as a suitable marker for adduct information.
Schema version is bumped to 3.7.2



Former-commit-id: 5c3c5191b92fbfdfc3d24321a1bac1c9a0d2a006

+++ pwiz_tools/BiblioSpec/src/SQTreader.cpp
@@ -143,7 +143,7 @@ void SQTreader::openRead(bool warnIfNotPercolated)
             if( sqtVersion->generatorName() == "Comet" ||
                     (sqtVersion->generatorName() == "Sequest" && *sqtVersion > SequestVersion("2.7"))){
                 string residue(1, modLetter);
-                float residueMass = getPeptideMass(residue, masses_);
+                float residueMass = static_cast<float>(getPeptideMass(residue, masses_));
                 modValue = modValue - residueMass;
             } 
             staticMods[(int)modLetter]=modValue;
